[test:backend/test-cov] - adding test coverage and marks to run functionalities tests separately
