+++ .gitignore
@@ -58,3 +58,6 @@ users.txt
 .env
 .pytest_cache
 .ruff_cache
+.coverage
+
+htmlcov
