+++ server/pyproject.toml
@@ -25,6 +25,18 @@ dev = [
     "ruff>=0.11.8",
 ]
 
+[tool.pytest.ini_options]
+markers = [
+    "auth",
+    "account",
+    "transaction",
+]
+
+filterwarnings = [
+    "ignore::DeprecationWarning",
+]
+
+
 [tool.ruff]
 exclude = [
     ".bzr",
